Restructure cypress component test folders
