+++ cypress/component/admin/users/Status.cy.tsx
