+++ cypress/component/admin/users/drivers/DriverDetail.cy.tsx
@@ -1,5 +1,5 @@
 import { Provider } from "react-redux";
-import DriverDetail from "./DriverDetail";
+import DriverDetail from "../../../../../src/components/admin/users/drivers/DriverDetail";
 import { store } from "@/store/store";
 
 describe("DriverDetail Component", () => {
