+++ cypress/component/common/admin/shimmers/README.md
@@ -1,0 +1,1 @@
+Remove this when you put your own test!
