+++ cypress/component/landing-page/benefits-section/BenefitsImage.cy.tsx
