+++ cypress/component/landing-page/benefits-section/BenefitsSection.cy.tsx
